Introduced a new struct Cluster
Represents a cluster of seq-sim-search Hits (Blast-Hits),
clustering is done as before based on overlap with the query
and scaled bit-score.
A cluster has two fields: the hit identifiers (Vec) and
its score (F64). The score is computed as the arithmetic mean
of the max similarity( query, hit ) and the fraction of hits
contained in the cluster.