Clustering hits computes aligned_query_region
Each cluster is now aware of the region of the query sequence that is
covered by _all_ hits in that respective cluster. Note that this
region is stored as an Option of a two-tupel of u32. If hits
align to overlapping regions or their similarity relies on their
descritions entirely, the query sequence region covered by _all_
hits in that cluster can be None.