Main thread handles HRD generation
Using transmitter and receiver technology from std::sync::mpsc to
send to the main thread Query IDs as messages whenever all Blast
result data of a Query has been read from the input tables. This
should squeeze some more performance out, because the child threads
parse the tables, and the parent thread, which was very idle, now
handles the generation of HRDs. Of course, in doing so, the main
thread 'locks()' the AnnotionProcess and thus child threads, that
might want to insert_query must wait in any case. But a little
performance should be gained. However, on needs to make sure
that messages are send otherwise this creates an endless loop.
The other (master branch) version is a little more elegant...