Added --blacklist-regexs (-b)
A file with regular expressions (Rust syntax), one per line. Any match to any of these regular expressions causes sequence similarity      search result descriptions ('stitle' in Blast terminology) to be discarded from the prot-scriber annotation process. If multiple --seq-sim-table (-s)         args are provided make sure the --header (-e) args appear in the correct order, e.g. the first -e arg will be used for the first -s arg, the second -e        will be used for the second -s and so on. Set to 'default' to use the hard coded default. Note that this is an expert option.